Rename work-space to workspace
